requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-smart-selects==1.5.4
idna==2.8
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
twilio==6.35.4
urllib3==1.25.8
